// gestion_paiement_frontend/src/models/Agent.ts
export type Agent = {
  id: number;
  matricule: string;
  civilite: string;
  nom: string;
  prenoms: string;
  cin: string;
  adresse: string;
  sexe: string;
  telephone: string;
  dateEntree: string;
  direction: string;
  service: string;
  division: string;
  statut: string;
  contrat: string;
  salaire: number;
  actif: boolean;
};